[NCL-6705] Aggregate retry data from log-event-duration
log-event-duration will now send "END" messages with the additional
'mdc.processContextVariant' that indicates a sub-category of a stage
(aka a retry happened).

kafka-store now adds all the durations for a build id with the same
stage, but with different `processContextVariant`. It will show the
total time of the stage, including with retries.

A future commit will be created to expose the "raw" data with the retry
information as a REST endpoint.